Reorganizacion de componentes en carpetas
